backend_project/requirements.txt
fastapi
uvicorn
pydantic
pyjwt
elasticsearch
python-dotenv
pandas
hdfs
sqlalchemy
alembic
asyncpg
psycopg2-binary
